driver/decompress.requirements.txt
pyyaml
py7zr
